renaming npm file
