+++ backend/engine/plugins/lib/trivy_common/generate_npm_locks.py
